refactor: extract maestro builder
