+++ maestro/swift/Sources/Core/main.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 let options = parseArguments(CommandLine.arguments)
-let maestro = Maestro(options: options)
+let maestro = makeMaestro(from: options)
 try startServer(on: options.port, maestro: maestro)
